Move cluster check selection saga in dedicated saga file
